src/services/docling/requirements.txt
docling>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
python-dotenv>=1.0.0
pypdf>=3.17.0
pillow>=10.0.0
aiofiles>=23.2.1
httpx>=0.25.0